+++ src/backend/index.ts
@@ -1,3 +1,7 @@
+import { warnPackageDeprecation } from '../deprecation-warning.js'
+
+warnPackageDeprecation('backend')
+
 export * from './verify-auth.js';
 export * from './csrf.js';
 export * from './authorization.js';